Move config examples into example_test.go
